+++ seeder/__main__.py
@@ -1,38 +1,12 @@
-from dataclasses import dataclass
+from models import User, Commit
 import time
 import requests
-import faker
 import random
-
-@dataclass(init=True, repr=True)
-class User:
-    name: str
-    surname: str
-    nick: str
-    avatar_url: str
-
-@dataclass(init=True, repr=True)
-class Commit:
-    creator_id: str
-    date: str | None = None
-    lines_added: int = 0
-    lines_removed: int = 0
-    description: str = ""
 
 def check_error(response: requests.Response):
     if not response.ok:
         print(f"Error: {response.status_code} {response.reason}")
         raise Exception(response.reason)
-
-fake = faker.Faker()
-
-def user_factory():
-    return User(
-        nick=fake.user_name(),
-        name=fake.first_name(),
-        surname=fake.last_name(),
-        avatar_url=f"https://picsum.photos/seed/{fake.uuid4()}/200/200"
-        )
 
 def get_random_user_id()->str|None:
     response: requests.Response = requests.get(USER_URL, headers=HEADER)
@@ -42,15 +16,6 @@ def get_random_user_id()->str|None:
         return None
     return random.choice(users)['userID']
 
-def commit_factory(creator_id: str):
-    return Commit(
-        creator_id=creator_id,
-        date=fake.date_time_between(start_date="-1y", end_date="now").isoformat(),
-        lines_added=random.randint(0, 100),
-        lines_removed=random.randint(0, 100),
-        description=fake.text(max_nb_chars=20)
-        )
-
 # Set the API endpoint URL and the data to send
 API_URL = 'http://127.0.0.1:5000'
 USER_URL = f'{API_URL}/user'
@@ -59,8 +24,8 @@ COMMIT_URL = f'{API_URL}/commit'
 # Set the API key as a header
 HEADER = {'x-access-token': 'dev'}
 
-r_user_factory = lambda: requests.put(USER_URL, json=user_factory().__dict__, headers=HEADER)
-r_commit_factory = lambda creator_id: requests.put(COMMIT_URL, json=commit_factory(creator_id).__dict__, headers=HEADER)
+r_user_factory = lambda: requests.put(USER_URL, json=User.fake().__dict__, headers=HEADER)
+r_commit_factory = lambda creator_id: requests.put(COMMIT_URL, json=Commit.fake(creator_id).__dict__, headers=HEADER)
 
 def main():
     print("Sending user data...")
